bring Auth model into User model; change how users are managed in database;
